+++ Changelog.md
@@ -10,6 +10,9 @@
     * [Pull request 162](https://bitbucket.org/osrf/vrx/pull-requests/162)
     
 
+1. Ocean reflection and refraction visual effects:
+    * [Pull request 165](https://bitbucket.org/osrf/vrx/pull-requests/165)
+
 ### VRX 1.1
 
 1. GUI overlay plugin (prototype):